When we're detyping a method application of the for expr.method(args) "expr.method" is represented in the AST as a field selection, but we don't want to transform that into a reflective field lookup when we call super.visitApply() so we have a chec kin visitSelect() that prevents the transformation of the Apply.meth node.
However if expr *itself* is a field selection (foo.bar.method()) then we *do*
want to transform the lower field selection node(s). Annoyingly this means we
have to do this manually in visitApply().

Alternatively I could come up with some complicated check to indicate that
Apply.meth.Select is not to be transformed but
Apply.meth.Select.Select[.Select...] is to be transformed.