Scratch that last comment. We can't just blindly translate the selected expression in an apply because we may have already translated it in super, if it wasn't itself a field selection (for example, if it was an application, like in: foo.bar().baz(), the selected expression would be foo.bar() which *would* get translated).
This forces me to do the "right" thing and figure out in visitSelect() whether
or not we're looking at the field selection that represents a method
application or a "real" field selection. This turned out to be easier than I
thought. If the path ends with Apply.meth then we don't want to translate,
otherwise we do. Score one for the awesome power of path expressions.