Handle the mapping of formal to actual type parameter arguments when determining the types of super method call arguments when casting said arguments back to their formal types. Since that can't have possibly made any sense, I'll explain with an example. Say you have:
class MyList<T> extends ArrayList<T> {
    public void add (T elem) { super.add(elem); }
}

ArrayList happens to be defined as:

class ArrayList<A> {
    public void add (A elem) { ... }
}

So when we get the Type object for ArrayList.add, it is expressed in terms of
the type variable A. However, we need to map that formal type argument to the
actual type argument T that MyList supplied when extending A in order to know
what the type of MyList.add looks like.

Fortunately, javac provides ClassType.supertype_field which contains a class's
supertype with formal type arguments mapped to actual type arguments. So we use
that in conjunction with Types.subst() which does type argument substitution to
turn ArrayList<A>.add(A) into MyList<T>.add(T), at which point we obtain
correct type names for the cast we have to insert in our detyped MyList.add
call. To wit:

class MyList<T> extends ArrayList<T> {
    public void add (T elem) {
        super.add(RT.<T>typeVarCast(Object.class, elem));
    }
}

If you're familiar with the code (meaning your initials are MDB) then you might
also notice that we're using typeVarCast() instead of checkedCast(). That's
because we cannot use the Class<C> provided to the checkedCast() method to
provide a return type C as that would be Object, not T. So we have an alternate
version of our dynamic cast which takes Class<?> and relies on the caller to
supply a return type in the form of an explicit type parameter.

I tried always using typeVarCast() but that caused the compiler to choke in
weird ways when it saw RT.<Integer>checkedCast(Integer.class, ...) which I
think may have something to do with boxing. So I restrict use of typeVarCast to
places where it is actually needed (which conveniently never involve boxable
types).

It's always fun when the length commit message exceeds the length of the code
it describes (not counting the new unit tests in this case).