We need to retain this annotation at runtime as we now must check (at runtime) whether a class was transformed to determine whether or not to pass dummies in type carrying arguments to its constructors (since we can't name mangle constructors to indicate that they were signature mangled like we can with methods).
Fortunately, if a class is transformed, then *all* of its constructors are
signature mangled, so we can make use of @Transformed in this way. The same
does not hold for its methods. A method may or may not be signature mangled
depending on whether it overrides a method in a non-transformed parent class or
interface. So for methods we require (simple) name mangling to tag methods as
signature mangled or not.