Oh the web of lies we weave when javac we do deceive. Normally method symbols are assigned when their enclosing class is complete()d, but for the classes that we fake-enter, we can't complete them because that commits things to javac's internal state that cause it to keel over later (at a minimum, it ends up synthesizing duplicate constructors for anonymous inner classes, and later when it goes to resolve said synthesized constructor it encounters overoad ambiguity and shoots itself in the head).
If we leave our methods with null symbols, then if an inner class method
encloses *yet another* inner class, the fake-entry of that inner class fails
because it looks at its enclosing scope (the method) and goes to investigate
its symbol and finds null. Antics (with the initials N.P.E.) ensue.

So we fake up method symbols when we encounter methods with no symbols, which
allows all of our fakery to proceed well enough without causing state changes
deep in the bowels of javac that doom us to later compilation failure. Since
I'm not creating real symbols for these methods, there are probably subtle
assumptions that happen later that will eventually fail when I determine that I
needed to properly set the static or protected flag or some other bit of
processing. Oh how excited I'll be when I see that crash and get to spend
another hour digging deeper into the bowels of javac.