Oh the plot, she is so thick! When we insert type carrying arguments directly into a call site (for calling a sigmangled super method or anonymous inner class constructor), we have to manually group up the varargs arguments into an array because we need to tack on type carrying arguments after them.
We also need to turn off the varargs flag on the value carrying argument,
otherwise it ends up with the type Object[] rather than Object, which then
causes a type failure when we pass in an array of the correct type
(e.g. int[]). When we're synthesizing an array to hold our varargs arguments,
it has to be the correct type, because it may eventually need to go to a
library method which will expect the correct type, and we don't want to break
correctly typed code. This means we have to insert casts back to the array
element type for any arguments that are grouped into this array, because
they'll have been detyped to object.

It's a twisty maze of passages!