Boggle. As is, sym.type for Comaparator<V>.compare(V,V) is resolved at the time that I discover that I need to cast the return value back to a valid type (as it's a library overrider). However sym.type for Callable<V>.call() is null.
Mind you, both of the MethodSymbols in this case were created by me, *with a
null type*. Someone is filling in the type of the former and not the latter. I
have no idea why.

Further boggle. If I uncomment "V foo = null", sym.type for Callable<V>.call()
magically gets filled in. When? By whom?! I must don my miner's helmet and dive
into the bowels of the javac code to find out.