TreeMaker.Type() generates an invalid AST when supplied with an unbounded wildcard (i.e. Class<?>). A wildcard expression is turned into an AST node that has a bound kind and an inner type. Valid instantions include:
- JCWildcard(SUPER, sometype)
- JCWildcard(EXTENDS, sometype)
- JCWildcard(UNBOUND, null)

TreeMaker type ends up generating:

- JCWildcard(UNBOUND, Object)

which causes javac to choke deep in the bowels of Flow and/or Check depending
on the phase of the moon. So rather than reimplement TreeMaker.Type() (which
does a lot of complicated fiddling), we just go in after the fact and ensure
that JCWildcard.inner is null for UNBOUNDed wildcards.

TreeMaker.Type() has other pesky bugs and is clearly something of a red headed
stepchild. It's used in a dozen or so places in the compiler internals, but
clearly none of those places tickle its bad behavior.