When we have an assignment to an ident or to this.ident we need to check whether the ident references a field defined in a library class. If so, we need to dynamically cast the RHS to the static type of the field because it will not have been detyped.
We opt to do this rather than do the assignment reflectively because we
currently rely on non-reflective field assignment to ensure definite
assignment. If we end up changing that, we could do all field assignment
reflectively, but I'm not sure that's actually more usable.

In this form, you find out about typos at compile time rather than runtime,
which I think is probably nice. On the flip side, if you have code that extends
a class and a field name gets changed or removed, then the derived class stops
compiling which may not be something you care about. The right tradeoff there
is certainly not obvious.