My original plan to "relax" interface use was to remove non-library interfaces from the implements declarations of application classes. However, interfaces may also have constant definitions and a common (though now considered deprecated) pattern is to implement an interface with no methods to allow its constant definitions to be used without qualification. Removing the interface from the implements clause of course breaks this.
The new plan is to strip out all method declarations from application
interfaces. Any constants remain available to implementors, but implementors do
not also inherit the burden of having to declare all of an interfaces methods.

This introduces yet another wrinkle, which is the use of the @Override
annotation on interface implementations. If we leave those in place for
application interface methods, the compiler will freak out when it goes to
check them later as the methods no longer "override" the interface
declarations. Since @Override is simply a compile-time convenience, I just
filter out all @Override annotations on detyped code.