Another test case for issue related to resolving static method when importing identically named interfaces
This requires two packages containing identically name interfaces which required a change to the Ant build.xml file in order to
build the two packages required by the test. This test demonstrates a class that implements an interface in one package and imports
another interface with the same name from a different package: this breaks static method resolution in a similar way to the
StaticMethodResolutionTest test case.