Oh the plot is always so thick. We can't just create an array to group the variable arguments, that would be too easy. It turns out that the varargs team and the generics team didn't coordinate early enough to realize that accepting parameterized varargs elements is tantamount to creating an array of parameterized types which is a no no (basically a hole in the type system).
So I can declare: <T> void foo (Function<T>... funcs) which under the hood is
actually doing: "new Function<C>[] { funcs }" (where C is some concrete type
dictated by the type of the funcs argument(s)) which is illegal in Java.

They actually generate a warning in Java 7 when you declare a function like
foo() above, since it's a bad idea. However, all of this means that I can't
simply resolve the type of the varargs array and insert a "new X[] { ... }"
expression where I want to. I have an idea of how I'll work around this (by
fighting fire with fire), but I wanted to commit this failing test with this
rant of a commit message just to ease my conscience.