Interesting, this test is failing for reasons other than the ones I expected. It's failing because the conversion of SneakyThrower.class.newInstance() into RT.invoke("newInstance" SneakyThrower.class) results in an attempt to call an inaccessible constructor. The constructor is accessible to sneakyThrow() but not accessible to RT.invoke().
Normally we deal with this by calling setAccessible() on the Method or
Constructor being called by RT, but in this case, we're calling a reflective
method on behalf of another class and have no idea that this method is a
reflective call or that it may suffer from accessibility problems. Oh the
twisty maze of passages.