These are the modifications needed to the Google Collection test suite to allow 44,760 of the test cases to pass. The changes generally involve two things:
1. Adjusting use of reflection to accommodate detyped methods.

2. Adding readObject() methods to serializable classes that contain transient
primitive fields which become non-primitive and are thus null after
unserialization (because no constructor is run when an object is unseriazized).

MapMakerTestSuite$ReferenceMapTest is currently disabled because it relies on
"new Integer(1) == new Integer(1)" not returning true. Presently RT.binop does
not distinguish between == where the operands have primitive static type (in
which case .equals() should be used) and where the operands have non-primitive
static type (in which case == should be used).