Completely avoid detyping constructors of classes that extend junit.framework.TestCase. This is a hack to work around the fact that JUnit expects such classes to have either a zero-argument constructor or a constructor that takes a single string argument (which becomes broken if detyped).
I would treat such constructors as library methods and detype their bodies, but
that introduces substantial complexity because we cannot insert detyped shadow
arguments before the call to super() which means that we would need to go in
and rewrite all instances of constructor arguments in any expressions inside
super() to their $T versions, which seems like a lot more trouble than it's
worth for this simple hack. So instead we just avoid detyping the constructors
entirely. Since the JUnit test case constructors that we're not detyping almost
invariably consist only of a call to "super(name)", this is not much of a
problem.