Unchecked exception compilation error under DuctileJ
This is a perplexing one: whether or not DuctileJ should support compilation of code without exception annotations on methods that originate exceptions we should at least be consistent. As is, this source file will compile (ant clean dtests) without reporting a compilation error. If, however, one deletes the source file NullPointerExceptionTest.java and recompiles then the error is reported.