Added another interface collision test case. In this case, we inherit "int size()" from HashMap, which means it remains undetyped. But we also implement an application-defined interface which declares an "int size()" method. That size() method gets detyped to "Object size()" and now our class no longer implements that interface.
When we strip methods out of interfaces, this is no problem, but stripping
methods causes the bridge method problem. I was thinking about not stripping
the methods out of interfaces and instead synthesizing fail-immediately methods
any methods that a class fails to implement for a given interface. This would
solve the bridge method problem, but it would run aground on this new problem.

Looks like it's back to the difficult solution for bridge methods. Sigh.